requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
django-cors-headers==4.2.0
msgpack==1.0.5
mysql-connector-python==8.1.0
protobuf==4.21.12
PyMySQL==1.1.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
channels-redis==3.4.0
Django==4.0.5
djangochannelsrestframework==1.0.0
djangorestframework==3.13.1
